Added /STER/pythonrepo/ path to config.py
An additional path was appended to the data directories list in
config.py . This new /STER/pythonrepo/ path will substitute the other
user paths (e.g. petered, kristofs, juries, etc) as the go-to place for
reference data, called by the scripts of the IvS python repository.
For now the user paths are left as-were, but users of the repository
are encouraged to change their paths to the new /STER/pythonrepo/
directory.