SED|INOUTS|CONF: make SED work with new MARCS models + remove z as var
Currently the model dir potins to /STER/ of Ana Escorza, but this sould
be changed in the future.
The possibility to exclude variables (e.g. metalicity) can be done more
general/as an instance attribute compared to as param in **kwargs that
is passed on from function to function. The plotting fixes are commented
out for the metalicity. This is not optimal...